% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_learnr.R
\name{question_is_correct.always_correct}
\alias{question_is_correct.always_correct}
\title{Title}
\usage{
question_is_correct.always_correct(question, value, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
