% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_learnr.R
\name{question_ui_initialize.sortable_rank}
\alias{question_ui_initialize.sortable_rank}
\title{Title}
\usage{
question_ui_initialize.sortable_rank(question, value, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
rank_ex <- sortable::question_rank(
"Sort the first 5 letters",
learnr::answer(LETTERS[1:5], correct = TRUE),
learnr::answer(rev(LETTERS[1:5]), correct = FALSE,
"Other direction!")
)
rank_ex
}
