% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_learnr.R
\name{question_ui_initialize.text_block}
\alias{question_ui_initialize.text_block}
\title{Title}
\usage{
question_ui_initialize.text_block(question, value, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
